fixing rating

+++ src/main/scala/models/User.scala
@@ -4,6 +4,6 @@ case class User(
   val userId: Int = 0 ,
   val nom : String,
   val vehicule : String = null,
-  val note : Int = 0,
+  val note : Double = 0.0,
   val nombreNote : Int = 0
 )
